+++ malcom/performers/management/commands/search_youtube_songs.py
@@ -1,7 +1,7 @@
+from commons.youtube_search import search_and_create_performer_songs
 from django.core.management.base import BaseCommand, CommandParser
 
 from performers.models import Performer, PerformerSocialLink
-from performers.youtube_search import search_and_create_performer_songs
 
 
 class Command(BaseCommand):
