+++ malcom/performers/management/commands/validate_youtube_sociallinks.py
@@ -4,9 +4,9 @@ import re
 import time
 from pathlib import Path
 
+from commons.youtube_utils import get_authorized_youtube_client
 from django.core.management.base import BaseCommand, CommandParser
 from django.utils import timezone
-from houses.youtube_utils import get_authorized_youtube_client
 
 from performers.models import PerformerSocialLink
 from performers.normalization import channel_name_matches
